Pair parents with their nearest frequency peer
We theorized that random parent seletion was *too* random, as this space
is kinda categorical in that the linear interpolation between two random
frequencies has effectively uniform probability of fitness. Not ideal,
but without any better ideas, we pair parents that are 'close' to
eachother. In theory, this helps the interpolation be more convex, but
it might also just make the interpolation effectively a no-op.