Improve(?) fitness computation
So the old fitness was problematic because individuals could achieve really
high fitness by matching exactly with background noise. This could be
fine if every frequency in the input 'gets' at least one individual, but
that is hard to guarantee. Instead, we really want to focus on the most
prominent frequencies in the input.

So, new fitness is the *amplitude of the nearest-neighbour frequency* in the
input. Seems better, but introduces the problem that the amplitude of
the individual has no effect on fitness; an individual matched perfectly to
a strong frequency is 'fit' even if it is inaudible. Not idea, but
unsure how to fix.

I think ideally we do something like expectation of individuals treated
as latent variables in a mixture model, but that would bump us up a
complexity class...