Add more test audio
I think maybe the issue we are running up against is that the EA process
is too random when an input audio is played back this fast.

c7c8.wav is literally just a 1-frequency slide, and this EA tracks it,
but not tightly. Mutation-rate/crossover-intensity drops could reign this in
maybe, but I'm unsure.

Theres also the problem that, as the slide moves, the output gets so
loud it distorts. No idea why this is happening. Mutation towards high
frequency ??